Move Switch, Route, and Link functionality from BeachCard to BeachesContainer
As it turns out, it is possible to use Routes, Switch, etc. without using a React hook.
It also looks better this and smoother this way:
By clicking the 'View Beach' button, you now go from viewing the BeachCard to viewing the Beach, and the button disappears.
Previously, clicking the 'View Beach' button would render the Beach AND the BeachCard.