Add Location info to BeachCard component, accounting for edge case
The edge case is that when the user refreshes the 'Saved Beaches' page, the Location info is initially undefined.
However, that is not the case when the user first starts the app - or refreshes from any other page - and navigates to 'Saved Beaches'.
The BeachesContainer seems to be re-rendering more than it ought to, but at this time I don't know why.