CORRECTLY memoize the selection of attractions by beach, and follow Separation of Concerns
This is similar to the previous commit for journal entries by beach.
In this case, I am also better following the Separation of Concerns principle:
I moved the selection of attractions by beach, from the BeachesContainer to the Beach component.
I could move that to the Attractions component later on, but as it stands now, Attractions is decoupled from the Beach component.