Move the selection of attractions by beach, from the Beach component to Attractions
I am moving this as far down the component tree as possible.
This way, when the Attractions component is (eventually) updated and re-rendered, the rest of the Beach component is not.
By the same token, this should prevent the Attractions component from re-rendering when the Beach is updated.
This also better obeys the Separation of Concerns.