Move the responsibility for selecting a location from Redux state, to the Location component
In order for all of this to work without any breaking changes, I had to update the selectors file and multiple components.
The goal is to limit re-renders of the Location component as much as possible.
By the same token, this should also limit re-renders of other components when a Location is changed.