Refactor reducer action for adding a location to state
Instead of creating an array of the state's values (which are all of the Location objects),
and checking to see whether that array includes the dispatched action's Location,
the reducer now checks whether the state has a corresponding value for the action's Location's id.
This is much easier and should make the Beach Journal load faster.