Bug fix: Update fetchBeaches action and the reducers by using only one dispatch
This prevents a bug/error message when refreshing the beaches page.
Up until now, all of my components were trying to render at the completion of the first dispatched action, before all of the data had been loaded into state.
This didn't happen when I first loaded the app because the home page didn't render any components that relied on the Redux state.
Redux thunk functions that are called with a dispatch are better/less bug prone if the dispatch is used once.
For this to work, the reducers had to use the same dispatched action when loading saved data.